fixed some GCC warnings

+++ test/test_bdecode.cpp
@@ -732,7 +732,7 @@ TORRENT_TEST(parse_length_overflow)
 		"d1:a9205357638345293824:11111",
 	};
 
-	for (int i = 0; i < sizeof(b)/sizeof(b[0]); ++i)
+	for (int i = 0; i < int(sizeof(b)/sizeof(b[0])); ++i)
 	{
 		error_code ec;
 		bdecode_node e;
